Tried to fix deletion of product

+++ backend/src/main/java/no/ntnu/xxs/product/ProductService.java
@@ -1,5 +1,6 @@
 package no.ntnu.xxs.product;
 
+import no.ntnu.xxs.exception.ProductAlreadyExistException;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -26,7 +27,7 @@ public class ProductService {
      * @param id the id of the product to get
      * @return an item with id as param
      */
-    public Product getProductById(long id) {
+    public Product getProductById(Long id) {
         Product productFound = null;
         Optional<Product> result = this.productRepository.findById(id);
         if (result.isPresent()) {
@@ -35,13 +36,33 @@ public class ProductService {
         return productFound;
     }
 
-    public void addProduct(Product product) {
-        Product productToAdd = new Product();
-        BeanUtils.copyProperties(product, productToAdd);
-        this.productRepository.save(productToAdd);
+    /**
+     * Adds a product to the database. If the product already exists,
+     * an exception is thrown.
+     * @param product the product to add.
+     * @throws ProductAlreadyExistException if the product already exists
+     */
+    public void addProduct(Product product) throws ProductAlreadyExistException {
+        Optional<Product> result = this.productRepository.findById(product.getId());
+        if (result.isEmpty()) {
+            this.productRepository.save(product);
+        } else {
+            throw new ProductAlreadyExistException("Product already in database");
+        }
     }
 
-    public void deleteProduct(long id) {
-        this.productRepository.deleteById(id);
+    /**
+     * Tries to delete a product from the repository
+     * @param id the id of the product to delete
+     * @return {@code true} if product is deleted, {@code false} otherwise
+     * // TODO: fix deletion of product added directly from dummy-data-initializer
+     */
+    public boolean deleteProduct(Long id) {
+        boolean deleted = false;
+        if (this.getProductById(id) != null) {
+            this.productRepository.deleteById(id);
+            deleted = true;
+        }
+        return deleted;
     }
 }
